refactor: reorganized fs logic into dir.go and file.go in /internal/fs
